Replicate rust-payjoin directory structure
Add bitcoin, ohttp, request modules. Remove types module.
